change: move components to separate folders
